Move PlaceOfferTab to root components Rename PlaceOfferTab => PlaceOfferSection
