+++ src/components/PlaceOfferSection/PlaceOfferContent/PlaceOfferContent.module.less
@@ -1,32 +1,57 @@
-.content {
-  padding: 24px;
+.summary {
+  border: 1px dotted var(--bg-border);
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 16px;
+  margin: 24px 0;
 
   @media (max-width: 640px) {
-    padding: 16px;
+    gap: 4px;
+    margin: 16px 0;
   }
 }
 
-.offerHeaderContent {
+.editSummary {
+  border-top: var(--border-less-primary);
   display: flex;
   align-items: center;
+  justify-content: space-around;
+  padding-top: 8px;
   gap: 16px;
-  margin-bottom: 8px;
 
-  @media (max-width: 640px) {
-    margin-bottom: 0;
-  }
-}
-.offerHeaderTitle {
-  font: var(--title-sm);
-  text-transform: uppercase;
-  line-height: 38px;
-
-  @media (max-width: 640px) {
-    font: var(--title-xs);
-    line-height: 32px;
+  & > div {
+    flex-direction: column-reverse;
+    align-items: center;
+    gap: 0;
   }
 }
 
+.fields {
+  display: flex;
+  align-items: center;
+  gap: 24px;
+  width: 100%;
+
+  & > div {
+    max-width: 216px;
+    width: 100%;
+  }
+
+  @media (max-width: 640px) {
+    gap: 16px;
+  }
+}
+
+.numericField {
+  max-width: unset !important;
+}
+
+.messageContainer {
+  height: 14px;
+}
+
+// OffetActionButtons
 .actionsButtonsContainer,
 .editModeButtonsContainer {
   display: flex;
@@ -55,13 +80,7 @@
   width: 100%;
 }
 
-.switchModeButtons {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.deleteOfferButton {
+.removeOfferButton {
   background-color: var(--additional-red-secondary) !important;
   border: 1px solid var(--additional-red-primary) !important;
 
@@ -69,26 +88,3 @@
     border: 1px solid var(--additional-red-primary-deep) !important;
   }
 }
-
-.switchButton {
-  height: 40px;
-
-  @media (max-width: 640px) {
-    height: 32px;
-  }
-
-  &.active {
-    color: var(--pure-black);
-    background: var(--accent-primary);
-  }
-
-  &.disabled {
-    background: var(--bg-secondary);
-    color: var(--content-secondary);
-  }
-}
-
-.borrowerMessage {
-  font: var(--body-text-sm);
-}
-
