+++ src/components/PlaceOfferSection/PlaceOfferContent/components/index.ts
@@ -1,0 +1,2 @@
+export * from './Summary'
+export * from './ActionsButtons'
