+++ src/components/PlaceOfferSection/PlaceOfferContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PlaceOfferContent'
