+++ src/components/PlaceOfferSection/PlaceOfferSection.module.less
@@ -1,0 +1,56 @@
+.content {
+  padding: 24px;
+
+  @media (max-width: 640px) {
+    padding: 16px;
+  }
+}
+
+.offerHeaderContent {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 8px;
+
+  @media (max-width: 640px) {
+    margin-bottom: 0;
+  }
+}
+.offerHeaderTitle {
+  font: var(--title-sm);
+  text-transform: uppercase;
+  line-height: 38px;
+
+  @media (max-width: 640px) {
+    font: var(--title-xs);
+    line-height: 32px;
+  }
+}
+
+.switchModeButtons {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.switchButton {
+  height: 40px;
+
+  @media (max-width: 640px) {
+    height: 32px;
+  }
+
+  &.active {
+    color: var(--pure-black);
+    background: var(--accent-primary);
+  }
+
+  &.disabled {
+    background: var(--bg-secondary);
+    color: var(--content-secondary);
+  }
+}
+
+.borrowerMessage {
+  font: var(--body-text-sm);
+}
