--- src/pages/LendPage/components/PlaceOfferTab/PlaceProOffer/PlaceProOffer.module.less
@@ -1,52 +1,0 @@
-.offerSummary {
-  border: 1px dotted var(--bg-border);
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  padding: 16px;
-  margin: 24px 0;
-
-  @media (max-width: 640px) {
-    gap: 4px;
-    margin: 16px 0;
-  }
-}
-
-.editOfferSummary {
-  border-top: var(--border-less-primary);
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-  padding-top: 8px;
-  gap: 16px;
-
-  & > div {
-    flex-direction: column-reverse;
-    align-items: center;
-    gap: 0;
-  }
-}
-
-.messageContainer {
-  height: 14px;
-}
-
-.fields {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-  width: 100%;
-
-  & > div {
-    max-width: 216px;
-    width: 100%;
-  }
-
-  @media (max-width: 640px) {
-    gap: 16px;
-  }
-}
-
-.numericField {
-  max-width: unset !important;
-}
