--- src/pages/LendPage/components/PlaceOfferTab/PlaceProOffer/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PlaceProOffer'
